Add conversation message image test
